+++ HISTORY.rst
@@ -4,6 +4,7 @@ Unreleased
 - Remove support for python 3.2.
 - Change json-body matcher to pass when the recorded request has the same
   content type as the issued request, and the content type is not json.
+- Update ``JSONBodyMatcher`` to work with requests 2.11+.
 
 0.3.0 - 2016-04-16
 ------------------